💄 Refactor mint related UI to <MintPageContainer>
